Rename `layers::service` to `layer`
